Refactor _CompressionRegistry
